Move all outcome components to a single folder PM-407
