+++ src/components/Outcome/OutcomeCategorical/index.js
