+++ src/components/Outcome/OutcomeCategorical/outcomeCategorical.scss
@@ -1,4 +1,4 @@
-@import '../../scss/vars.scss';
+@import '../../../scss/vars.scss';
 
 
 $outcome-bar-height: 15px;
