+++ src/components/Outcome/OutcomeScalar/index.js
