+++ src/components/Outcome/OutcomeScalar/outcomeScalar.scss
@@ -1,4 +1,4 @@
-@import '../../scss/vars.scss';
+@import '../../../scss/vars.scss';
 
 $expandable-arrow-size: 7px;
