+++ src/components/Outcome/WinningOutcome/WinningOutcome.scss
@@ -1,0 +1,33 @@
+.outcome {
+  display: flex;
+  margin: 10px 0 10px 3px;
+  align-items: center;
+  justify-content: flex-start;
+  text-transform: uppercase;
+  font-weight: 500;
+}
+
+.icon {
+  background: url('../../../assets/img/icons/icon_winningOutcome.svg');
+  background-size: contain;
+  width: 40px;
+  height: 40px;
+  margin-left: -10px;
+}
+
+.label {
+  color: #87949c;
+  font-size: 12px;
+}
+
+.textBox {
+  margin-left: 10px;
+  background-color: #87949c;
+  color: #fff;
+  border-radius: 6px;
+  padding: 5px 10px;
+}
+
+.unit {
+  text-transform: none;
+}
