+++ src/components/Outcome/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import OutcomeCategorical from 'components/OutcomeCategorical'
-import OutcomeScalar from 'components/OutcomeScalar'
-import WinningOutcome from 'components/WinningOutcome'
+import OutcomeCategorical from 'components/Outcome/OutcomeCategorical'
+import OutcomeScalar from 'components/Outcome/OutcomeScalar'
+import WinningOutcome from 'components/Outcome/WinningOutcome'
 import { OUTCOME_TYPES } from 'utils/constants'
 import { marketShape } from 'utils/shapes'
 import { isMarketResolved } from 'utils/helpers'
@@ -9,6 +9,7 @@ import PropTypes from 'prop-types'
 
 const Outcome = ({ market, opts = { showOnlyTrendingOutcome: false } }) => {
   const { event: { type: eventType } } = market
+  const showWinningOutcome = isMarketResolved(market)
 
   let outcomeComponent = eventType === OUTCOME_TYPES.CATEGORICAL ? (
     <OutcomeCategorical market={market} opts={opts} />
@@ -16,7 +17,7 @@ const Outcome = ({ market, opts = { showOnlyTrendingOutcome: false } }) => {
     <OutcomeScalar market={market} opts={opts} />
   )
 
-  if (isMarketResolved(market)) {
+  if (showWinningOutcome) {
     outcomeComponent = <WinningOutcome market={market} />
   }
