--- src/components/WinningOutcome/WinningOutcome.scss
@@ -1,35 +1,0 @@
-.outcome__winning {
-  &--container {
-    display: flex;
-    margin: 10px 0 10px 3px;
-    align-items: center;
-    justify-content: flex-start;
-    text-transform: uppercase;
-    font-weight: 500;
-  }
-
-  &--icon {
-    background: url("../../assets/img/icons/icon_winningOutcome.svg");
-    background-size: contain;
-    width: 40px;
-    height: 40px;
-    margin-left: -10px;
-  }
-
-  &--label {
-    color: #87949c;
-    font-size: 12px;
-  }
-
-  &--box {
-    margin-left: 10px;
-    background-color: #87949c;
-    color: #fff;
-    border-radius: 6px;
-    padding: 5px 10px;
-  }
-
-  &--unit {
-      text-transform: none;
-  }
-}
