+++ webpack.dev.config.js
@@ -60,6 +60,8 @@ module.exports = {
             loader: 'css-loader',
             options: {
               sourceMap: true,
+              importLoaders: 2,
+              localIdentName: '[name]--[local]--[hash:base64:8]',
             },
           },
           {
@@ -68,7 +70,10 @@ module.exports = {
               sourceMap: true,
             },
           },
-          { loader: 'sass-loader', options: { sourceMap: true } },
+          {
+            loader: 'sass-loader',
+            options: { sourceMap: true, localIdentName: '[path][name]__[local]--[hash:base64:5]' },
+          },
         ],
       },
       {
